New forget endpoint and support for pre-trained model init control.
This commit brings the implementation for the new forget endpoint, which in turn re=initialise the current model and generates new predictions for all images (nodes) in the current board.

An extra (query-string) parameter is also added to main `faces` endpoint to allow controlling whether learning machine model should be returned pre-trained or not!