New EvaluationDataSource to be used for metric computation
A new instance of `EvaluationDataSource` is now available to be used specifically for metric calculation (i.e. Accuracy).
The new special data source includes a list of selected samples per each class (10 each) that will be consistently used *only* to evaluate the online performance of the model (i.e. the learning machine).