New set of notebooks with Exploratory Analysis for (Evaluation) sample selection
This notebook contains all the code, and the analysis performed to identify and select the candidate faces (from the Validation Set) to be used for metric computation.
The (data-driven) strategy considers an AutoEncoder trained on the training set, in an unsupervised fashion.
The generated representation in the latent space is then passed on to a UMAP Dimension reduction model, to plot samples onto the embedding space (one for each data partition).

For each generated embedding, centroids per each class are generated and the 10 closest samples to the centroids are identified.
These are the chosen candidate samples to be used in metric evaluation.