New FERPlus dataset added.
FERPlus Dataset now is fully integrated into the backend as an available source.

FERPlus is now available as a full-fledge dataset (to be used with the frontend) as well as
plugged-in for evaluation as metric dataset.
Corresponding functions and settings have been updated accordingly.

Note: samples selection for FERPlus **Metric** Dataset have been chosen using a similar
approach used for FER (further details in the next commit with the attached notebook)